+++ src/backend/joanie/tests/core/test_commands_generate_certificates.py
@@ -164,6 +164,7 @@ class CreateCertificatesTestCase(TestCase):
         with self.assertNumQueries(16):
             call_command("generate_certificates", product=product_1.id)
         self.assertEqual(certificate_qs.filter(order=orders[0]).count(), 1)
+        self.assertEqual(certificate_qs.filter(order=orders[1]).count(), 0)
 
         # Then a certificate should be generated for the 2nd product
         with self.assertNumQueries(16):
@@ -266,6 +267,7 @@ class CreateCertificatesTestCase(TestCase):
         with self.assertNumQueries(16):
             call_command("generate_certificates", product=product_1.id)
         self.assertEqual(certificate_qs.filter(order=orders[0]).count(), 1)
+        self.assertEqual(certificate_qs.filter(order=orders[1]).count(), 0)
 
         # Then a certificate should be generated for the 2nd product
         with self.assertNumQueries(16):